change confirmation

+++ public/build/js/app.js
@@ -17,15 +17,6 @@ app.directive('checkboxField', function() {
   };
 });
 
-app.directive('confirmModal', function() {
-  return {
-    restrict: 'AE',
-    templateUrl: '/views/directives/confirm_modal.html',
-    scope: true,
-    link: function(scope, element, attrs) {}
-  };
-});
-
 app.directive('datetimepicker', function() {
   return {
     restrict: 'AE',
@@ -96,45 +87,6 @@ app.directive('radioField', function() {
   };
 });
 
-app.controller('confirmDeleteStoreCtrl', function($scope, $rootScope, $uibModalInstance) {
-  $scope.modalTitle = 'Delete store';
-  $scope.modalText = 'Are you sure?';
-  $scope.ok = function() {
-    $uibModalInstance.close();
-  };
-  return $scope.cancel = function() {
-    $uibModalInstance.dismiss();
-  };
-});
-
-app.controller('confirmDeleteUserCtrl', function($scope, $rootScope, $uibModalInstance) {
-  $scope.modalTitle = 'Delete user';
-  $scope.modalText = 'Are you sure?';
-  $scope.ok = function() {
-    $uibModalInstance.close();
-  };
-  return $scope.cancel = function() {
-    $uibModalInstance.dismiss();
-  };
-});
-
-app.controller('indexStoreCtrl', function($scope, $rootScope, $http, $uibModal) {
-  return $scope.deleteStore = function(id) {
-    $uibModal.open({
-      templateUrl: 'confirmModal.html',
-      controller: 'confirmDeleteStoreCtrl',
-      size: 'md'
-    }).result.then((function() {
-      $http({
-        method: 'DELETE',
-        url: '/stores/' + id
-      }).then((function(response) {
-        window.location.reload();
-      }), function(response) {});
-    }), function(res) {});
-  };
-});
-
 app.controller('createUserCtrl', [
   "$scope", "lodash", function($scope, lodash) {
     $scope.passInput = document.querySelector('.password-input');
@@ -156,17 +108,30 @@ app.controller('createUserCtrl', [
 
 app.controller('indexUserCtrl', function($scope, $rootScope, $http, $uibModal) {
   return $scope.deleteUser = function(id) {
-    $uibModal.open({
-      templateUrl: 'confirmModal.html',
-      controller: 'confirmDeleteUserCtrl',
-      size: 'md'
-    }).result.then((function() {
+    var confirmation;
+    confirmation = confirm('Are you sure?');
+    if (confirmation) {
       $http({
         method: 'DELETE',
         url: '/users/' + id
       }).then((function(response) {
         window.location.reload();
-      }), function(response) {});
-    }), function(res) {});
+      }));
+    }
+  };
+});
+
+app.controller('indexStoreCtrl', function($scope, $rootScope, $http, $uibModal) {
+  return $scope.deleteStore = function(id) {
+    var confirmation;
+    confirmation = confirm('Are you sure?');
+    if (confirmation) {
+      $http({
+        method: 'DELETE',
+        url: '/stores/' + id
+      }).then((function(response) {
+        window.location.reload();
+      }));
+    }
   };
 });
