+++ resources/assets/js/controllers/users/index_user_ctrl.coffee
@@ -1,19 +1,14 @@
 app.controller 'indexUserCtrl', ($scope, $rootScope, $http, $uibModal) ->
     $scope.deleteUser = (id) ->
-        $uibModal.open(
-            templateUrl: 'confirmModal.html'
-            controller: 'confirmDeleteUserCtrl'
-            size: 'md'
-        ).result.then (->
+        confirmation = confirm('Are you sure?')
+
+        if confirmation
             $http(
                 method: 'DELETE'
                 url: '/users/' + id).then ((response) ->
                     window.location.reload()
                     return
-            ), (response) ->
-                return
-            return
-        ), (res)->
-            return
+            )
+
         return
